requirements.txt
pyparrot==1.5.10
opencv-python==4.0.0.21
zeroconf==0.21.3
PyQt5==5.11.3
python-vlc==3.0.4106